Use One-way ANOVA test,Kruskal-Wallist test for null hypothesis
The test is summarized as follows:
Null Hypothesis: H0: µpacbo=µhi=µlow
Alternative Hypothesis: Ha: not H0 
p-value(high)= 0.4056 (>𝜶=0.05),
p-value(low)= 0.3181 (>𝜶=0.05),
p-value(placebo)=0.3982(>𝜶=0.05) 
decision rule: reject H0 if p-value< 𝜶=0.05
So in the normality test, we reject the null hypothesis H0 because all the p-value is great than 𝜶=0.05, so do not have sufficient evidence to say that there is a difference mean item number recalled by patients in the three-dose group.   
However, in the One-way ANOVA test and the Kruskal-Wallis test, we both cannot reject the null hypothesis H0, also the p-value in the two tests is very marginal. Therefore, we do not have enough evidence to claim that there is a difference in the average number of items recalled by subjects who are given oral therapeutic agents by three does group. As our data is not normal in our study, we choose the test for homogeneity of variance, we conclude that there is a difference in the average number of items recalled by subjects who are given oral therapeutic agents by three does groups.